requirements.txt
transformers
langchain
sentence-transformers
chromadb
pypdf
python-dotenv
torch
